Reorganize dataset make functions
