+++ photutils/datasets/__init__.py
@@ -4,5 +4,8 @@ This subpackage contains tools for making or loading datasets for
 examples and tests.
 """
 
+from .images import *  # noqa: F401, F403
 from .load import *  # noqa: F401, F403
-from .make import *  # noqa: F401, F403
+from .noise import *  # noqa: F401, F403
+from .sources import *  # noqa: F401, F403
+from .wcs import *  # noqa: F401, F403
