+++ photutils/datasets/images.py
@@ -1,374 +1,28 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """
-This module provides tools for making example datasets for examples and
-tests.
+This module provides tools for making simulated images for documentation
+examples and tests.
 """
 
 import math
 import warnings
 
-import astropy.units as u
 import numpy as np
-from astropy import coordinates as coord
 from astropy.convolution import discretize_model
-from astropy.io import fits
 from astropy.modeling import models
 from astropy.nddata import overlap_slices
 from astropy.table import QTable
 from astropy.utils.exceptions import AstropyUserWarning
-from astropy.wcs import WCS
 
 from photutils.psf import IntegratedGaussianPRF
 from photutils.utils._coords import make_random_xycoords
-from photutils.utils._misc import _get_meta
 from photutils.utils._parameters import as_pair
 from photutils.utils._progress_bars import add_progress_bar
 
-__all__ = ['apply_poisson_noise', 'make_noise_image',
-           'make_random_models_table', 'make_random_gaussians_table',
-           'make_model_sources_image', 'make_gaussian_sources_image',
+__all__ = ['make_model_sources_image', 'make_gaussian_sources_image',
            'make_4gaussians_image', 'make_100gaussians_image',
-           'make_wcs', 'make_gwcs', 'make_imagehdu',
            'make_gaussian_prf_sources_image',
            'make_test_psf_data']
-
-__doctest_requires__ = {('make_gwcs'): ['gwcs']}
-
-
-def apply_poisson_noise(data, seed=None):
-    """
-    Apply Poisson noise to an array, where the value of each element in
-    the input array represents the expected number of counts.
-
-    Each pixel in the output array is generated by drawing a random
-    sample from a Poisson distribution whose expectation value is given
-    by the pixel value in the input array.
-
-    Parameters
-    ----------
-    data : array_like
-        The array on which to apply Poisson noise.  Every pixel in the
-        array must have a positive value (i.e., counts).
-
-    seed : int, optional
-        A seed to initialize the `numpy.random.BitGenerator`. If `None`,
-        then fresh, unpredictable entropy will be pulled from the OS.
-
-    Returns
-    -------
-    result : `~numpy.ndarray`
-        The data array after applying Poisson noise.
-
-    See Also
-    --------
-    make_noise_image
-
-    Examples
-    --------
-    .. plot::
-        :include-source:
-
-        import matplotlib.pyplot as plt
-        from photutils.datasets import (apply_poisson_noise,
-                                        make_4gaussians_image)
-
-        data1 = make_4gaussians_image(noise=False)
-        data2 = apply_poisson_noise(data1, seed=0)
-
-        # plot the images
-        fig, (ax1, ax2) = plt.subplots(2, 1, figsize=(8, 8))
-        ax1.imshow(data1, origin='lower', interpolation='nearest')
-        ax1.set_title('Original image')
-        ax2.imshow(data2, origin='lower', interpolation='nearest')
-        ax2.set_title('Original image with Poisson noise applied')
-    """
-    data = np.asanyarray(data)
-    if np.any(data < 0):
-        raise ValueError('data must not contain any negative values')
-
-    rng = np.random.default_rng(seed)
-
-    return rng.poisson(data)
-
-
-def make_noise_image(shape, distribution='gaussian', mean=None, stddev=None,
-                     seed=None):
-    r"""
-    Make a noise image containing Gaussian or Poisson noise.
-
-    Parameters
-    ----------
-    shape : 2-tuple of int
-        The shape of the output 2D image.
-
-    distribution : {'gaussian', 'poisson'}
-        The distribution used to generate the random noise:
-
-            * ``'gaussian'``: Gaussian distributed noise.
-            * ``'poisson'``: Poisson distributed noise.
-
-    mean : float
-        The mean of the random distribution.  Required for both Gaussian
-        and Poisson noise.  The default is 0.
-
-    stddev : float, optional
-        The standard deviation of the Gaussian noise to add to the
-        output image.  Required for Gaussian noise and ignored for
-        Poisson noise (the variance of the Poisson distribution is equal
-        to its mean).
-
-    seed : int, optional
-        A seed to initialize the `numpy.random.BitGenerator`. If `None`,
-        then fresh, unpredictable entropy will be pulled from the OS.
-
-    Returns
-    -------
-    image : 2D `~numpy.ndarray`
-        Image containing random noise.
-
-    See Also
-    --------
-    apply_poisson_noise
-
-    Examples
-    --------
-    .. plot::
-        :include-source:
-
-        import matplotlib.pyplot as plt
-        from photutils.datasets import make_noise_image
-
-        # make Gaussian and Poisson noise images
-        shape = (100, 100)
-        image1 = make_noise_image(shape, distribution='gaussian', mean=0.,
-                                  stddev=5.)
-        image2 = make_noise_image(shape, distribution='poisson', mean=5.)
-
-        # plot the images
-        fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(8, 4))
-        ax1.imshow(image1, origin='lower', interpolation='nearest')
-        ax1.set_title(r'Gaussian noise ($\mu=0$, $\sigma=5.$)')
-        ax2.imshow(image2, origin='lower', interpolation='nearest')
-        ax2.set_title(r'Poisson noise ($\mu=5$)')
-    """
-    if mean is None:
-        raise ValueError('"mean" must be input')
-
-    rng = np.random.default_rng(seed)
-
-    if distribution == 'gaussian':
-        if stddev is None:
-            raise ValueError('"stddev" must be input for Gaussian noise')
-        image = rng.normal(loc=mean, scale=stddev, size=shape)
-    elif distribution == 'poisson':
-        image = rng.poisson(lam=mean, size=shape)
-    else:
-        raise ValueError(f'Invalid distribution: {distribution}. Use either '
-                         '"gaussian" or "poisson".')
-
-    return image
-
-
-def make_random_models_table(n_sources, param_ranges, seed=None):
-    """
-    Make a `~astropy.table.QTable` containing randomly generated
-    parameters for an Astropy model to simulate a set of sources.
-
-    Each row of the table corresponds to a source whose parameters are
-    defined by the column names.  The parameters are drawn from a
-    uniform distribution over the specified input ranges.
-
-    The output table can be input into :func:`make_model_sources_image`
-    to create an image containing the model sources.
-
-    Parameters
-    ----------
-    n_sources : float
-        The number of random model sources to generate.
-
-    param_ranges : dict
-        The lower and upper boundaries for each of the model parameters
-        as a dictionary mapping the parameter name to its ``(lower,
-        upper)`` bounds.
-
-    seed : int, optional
-        A seed to initialize the `numpy.random.BitGenerator`. If `None`,
-        then fresh, unpredictable entropy will be pulled from the OS.
-
-    Returns
-    -------
-    table : `~astropy.table.QTable`
-        A table of parameters for the randomly generated sources.  Each
-        row of the table corresponds to a source whose model parameters
-        are defined by the column names.  The column names will be the
-        keys of the dictionary ``param_ranges``.
-
-    See Also
-    --------
-    make_random_gaussians_table, make_model_sources_image
-
-    Notes
-    -----
-    To generate identical parameter values from separate function
-    calls, ``param_ranges`` must have the same parameter ranges and the
-    ``seed`` must be the same.
-
-    Examples
-    --------
-    >>> from photutils.datasets import make_random_models_table
-    >>> n_sources = 5
-    >>> param_ranges = {'amplitude': [500, 1000],
-    ...                 'x_mean': [0, 500],
-    ...                 'y_mean': [0, 300],
-    ...                 'x_stddev': [1, 5],
-    ...                 'y_stddev': [1, 5],
-    ...                 'theta': [0, np.pi]}
-    >>> sources = make_random_models_table(n_sources, param_ranges,
-    ...                                    seed=0)
-    >>> for col in sources.colnames:
-    ...     sources[col].info.format = '%.8g'  # for consistent table output
-    >>> print(sources)
-    amplitude   x_mean    y_mean    x_stddev  y_stddev   theta
-    --------- --------- ---------- --------- --------- ---------
-    818.48084 456.37779  244.75607 1.7026225 1.1132787 1.2053586
-    634.89336 303.31789 0.82155005 4.4527157 1.4971331 3.1328274
-    520.48676 364.74828  257.22128 3.1658449 3.6824977 3.0813851
-    508.26382  271.8125  10.075673 2.1988476  3.588758 2.1536937
-    906.63512 467.53621  218.89663 2.6907489 3.4615404 2.0434781
-    """
-    rng = np.random.default_rng(seed)
-
-    sources = QTable()
-    sources.meta.update(_get_meta())  # keep sources.meta type
-    for param_name, (lower, upper) in param_ranges.items():
-        # Generate a column for every item in param_ranges, even if it
-        # is not in the model (e.g., flux). However, such columns will be
-        # ignored when rendering the image.
-        sources[param_name] = rng.uniform(lower, upper, n_sources)
-
-    return sources
-
-
-def make_random_gaussians_table(n_sources, param_ranges, seed=None):
-    """
-    Make a `~astropy.table.QTable` containing randomly generated
-    parameters for 2D Gaussian sources.
-
-    Each row of the table corresponds to a Gaussian source whose
-    parameters are defined by the column names.  The parameters are
-    drawn from a uniform distribution over the specified input ranges.
-
-    The output table can be input into
-    :func:`make_gaussian_sources_image` to create an image containing
-    the 2D Gaussian sources.
-
-    Parameters
-    ----------
-    n_sources : float
-        The number of random Gaussian sources to generate.
-
-    param_ranges : dict
-        The lower and upper boundaries for each of the
-        `~astropy.modeling.functional_models.Gaussian2D` parameters
-        as a dictionary mapping the parameter name to its ``(lower,
-        upper)`` bounds. The dictionary keys must be valid
-        `~astropy.modeling.functional_models.Gaussian2D` parameter
-        names or ``'flux'``. If ``'flux'`` is specified, but not
-        ``'amplitude'`` then the 2D Gaussian amplitudes will be
-        calculated and placed in the output table. If both ``'flux'``
-        and ``'amplitude'`` are specified, then ``'flux'`` will be
-        ignored. Model parameters not defined in ``param_ranges`` will
-        be set to the default value.
-
-    seed : int, optional
-        A seed to initialize the `numpy.random.BitGenerator`. If `None`,
-        then fresh, unpredictable entropy will be pulled from the OS.
-
-    Returns
-    -------
-    table : `~astropy.table.QTable`
-        A table of parameters for the randomly generated Gaussian
-        sources.  Each row of the table corresponds to a Gaussian source
-        whose parameters are defined by the column names.
-
-    See Also
-    --------
-    make_random_models_table, make_gaussian_sources_image
-
-    Notes
-    -----
-    To generate identical parameter values from separate function
-    calls, ``param_ranges`` must have the same parameter ranges and the
-    ``seed`` must be the same.
-
-    Examples
-    --------
-    >>> from photutils.datasets import make_random_gaussians_table
-    >>> n_sources = 5
-    >>> param_ranges = {'amplitude': [500, 1000],
-    ...                 'x_mean': [0, 500],
-    ...                 'y_mean': [0, 300],
-    ...                 'x_stddev': [1, 5],
-    ...                 'y_stddev': [1, 5],
-    ...                 'theta': [0, np.pi]}
-    >>> sources = make_random_gaussians_table(n_sources, param_ranges,
-    ...                                       seed=0)
-    >>> for col in sources.colnames:
-    ...     sources[col].info.format = '%.8g'  # for consistent table output
-    >>> print(sources)
-    amplitude   x_mean    y_mean    x_stddev  y_stddev   theta
-    --------- --------- ---------- --------- --------- ---------
-    818.48084 456.37779  244.75607 1.7026225 1.1132787 1.2053586
-    634.89336 303.31789 0.82155005 4.4527157 1.4971331 3.1328274
-    520.48676 364.74828  257.22128 3.1658449 3.6824977 3.0813851
-    508.26382  271.8125  10.075673 2.1988476  3.588758 2.1536937
-    906.63512 467.53621  218.89663 2.6907489 3.4615404 2.0434781
-
-    To specifying the flux range instead of the amplitude range:
-
-    >>> param_ranges = {'flux': [500, 1000],
-    ...                 'x_mean': [0, 500],
-    ...                 'y_mean': [0, 300],
-    ...                 'x_stddev': [1, 5],
-    ...                 'y_stddev': [1, 5],
-    ...                 'theta': [0, np.pi]}
-    >>> sources = make_random_gaussians_table(n_sources, param_ranges,
-    ...                                       seed=0)
-    >>> for col in sources.colnames:
-    ...     sources[col].info.format = '%.8g'  # for consistent table output
-    >>> print(sources)
-       flux     x_mean    y_mean    x_stddev  y_stddev   theta   amplitude
-    --------- --------- ---------- --------- --------- --------- ---------
-    818.48084 456.37779  244.75607 1.7026225 1.1132787 1.2053586 68.723678
-    634.89336 303.31789 0.82155005 4.4527157 1.4971331 3.1328274 15.157778
-    520.48676 364.74828  257.22128 3.1658449 3.6824977 3.0813851 7.1055501
-    508.26382  271.8125  10.075673 2.1988476  3.588758 2.1536937 10.251089
-    906.63512 467.53621  218.89663 2.6907489 3.4615404 2.0434781 15.492093
-
-    Note that in this case the output table contains both a flux and
-    amplitude column.  The flux column will be ignored when generating
-    an image of the models using :func:`make_gaussian_sources_image`.
-    """
-    sources = make_random_models_table(n_sources, param_ranges,
-                                       seed=seed)
-
-    # convert Gaussian2D flux to amplitude
-    if 'flux' in param_ranges and 'amplitude' not in param_ranges:
-        model = models.Gaussian2D(x_stddev=1, y_stddev=1)
-
-        if 'x_stddev' in sources.colnames:
-            xstd = sources['x_stddev']
-        else:
-            xstd = model.x_stddev.value  # default
-        if 'y_stddev' in sources.colnames:
-            ystd = sources['y_stddev']
-        else:
-            ystd = model.y_stddev.value  # default
-
-        sources = sources.copy()
-        sources['amplitude'] = sources['flux'] / (2.0 * np.pi * xstd * ystd)
-
-    return sources
 
 
 def make_model_sources_image(shape, model, source_table, oversample=1):
@@ -768,191 +422,6 @@ def make_100gaussians_image(noise=True):
     return data
 
 
-def make_wcs(shape, galactic=False):
-    """
-    Create a simple celestial `~astropy.wcs.WCS` object in either the
-    ICRS or Galactic coordinate frame.
-
-    Parameters
-    ----------
-    shape : 2-tuple of int
-        The shape of the 2D array to be used with the output
-        `~astropy.wcs.WCS` object.
-
-    galactic : bool, optional
-        If `True`, then the output WCS will be in the Galactic
-        coordinate frame.  If `False` (default), then the output WCS
-        will be in the ICRS coordinate frame.
-
-    Returns
-    -------
-    wcs : `astropy.wcs.WCS` object
-        The world coordinate system (WCS) transformation.
-
-    See Also
-    --------
-    make_gwcs, make_imagehdu
-
-    Notes
-    -----
-    The `make_gwcs` function returns an equivalent WCS transformation to
-    this one, but as a `gwcs.wcs.WCS` object.
-
-    Examples
-    --------
-    >>> from photutils.datasets import make_wcs
-    >>> shape = (100, 100)
-    >>> wcs = make_wcs(shape)
-    >>> print(wcs.wcs.crpix)  # doctest: +FLOAT_CMP
-    [50. 50.]
-    >>> print(wcs.wcs.crval)  # doctest: +FLOAT_CMP
-    [197.8925      -1.36555556]
-    """
-    wcs = WCS(naxis=2)
-    rho = np.pi / 3.0
-    scale = 0.1 / 3600.0  # 0.1 arcsec/pixel in deg/pix
-
-    wcs.pixel_shape = shape
-    wcs.wcs.crpix = [shape[1] / 2, shape[0] / 2]  # 1-indexed (x, y)
-    wcs.wcs.crval = [197.8925, -1.36555556]
-    wcs.wcs.cunit = ['deg', 'deg']
-    wcs.wcs.cd = [[-scale * np.cos(rho), scale * np.sin(rho)],
-                  [scale * np.sin(rho), scale * np.cos(rho)]]
-    if not galactic:
-        wcs.wcs.radesys = 'ICRS'
-        wcs.wcs.ctype = ['RA---TAN', 'DEC--TAN']
-    else:
-        wcs.wcs.ctype = ['GLON-CAR', 'GLAT-CAR']
-
-    return wcs
-
-
-def make_gwcs(shape, galactic=False):
-    """
-    Create a simple celestial gWCS object in the ICRS coordinate frame.
-
-    This function requires the `gwcs
-    <https://github.com/spacetelescope/gwcs>`_ package.
-
-    Parameters
-    ----------
-    shape : 2-tuple of int
-        The shape of the 2D array to be used with the output
-        `~gwcs.wcs.WCS` object.
-
-    galactic : bool, optional
-        If `True`, then the output WCS will be in the Galactic
-        coordinate frame.  If `False` (default), then the output WCS
-        will be in the ICRS coordinate frame.
-
-    Returns
-    -------
-    wcs : `gwcs.wcs.WCS` object
-        The generalized world coordinate system (WCS) transformation.
-
-    See Also
-    --------
-    make_wcs, make_imagehdu
-
-    Notes
-    -----
-    The `make_wcs` function returns an equivalent WCS transformation to
-    this one, but as an `astropy.wcs.WCS` object.
-
-    Examples
-    --------
-    >>> from photutils.datasets import make_gwcs
-    >>> shape = (100, 100)
-    >>> gwcs = make_gwcs(shape)
-    >>> print(gwcs)
-      From      Transform
-    -------- ----------------
-    detector linear_transform
-        icrs             None
-    """
-    from gwcs import coordinate_frames as cf
-    from gwcs import wcs as gwcs_wcs
-
-    rho = np.pi / 3.0
-    scale = 0.1 / 3600.0  # 0.1 arcsec/pixel in deg/pix
-
-    shift_by_crpix = (models.Shift((-shape[1] / 2) + 1)
-                      & models.Shift((-shape[0] / 2) + 1))
-
-    cd_matrix = np.array([[-scale * np.cos(rho), scale * np.sin(rho)],
-                          [scale * np.sin(rho), scale * np.cos(rho)]])
-
-    rotation = models.AffineTransformation2D(cd_matrix, translation=[0, 0])
-    rotation.inverse = models.AffineTransformation2D(
-        np.linalg.inv(cd_matrix), translation=[0, 0])
-
-    tan = models.Pix2Sky_TAN()
-    celestial_rotation = models.RotateNative2Celestial(197.8925, -1.36555556,
-                                                       180.0)
-
-    det2sky = shift_by_crpix | rotation | tan | celestial_rotation
-    det2sky.name = 'linear_transform'
-
-    detector_frame = cf.Frame2D(name='detector', axes_names=('x', 'y'),
-                                unit=(u.pix, u.pix))
-
-    if galactic:
-        sky_frame = cf.CelestialFrame(reference_frame=coord.Galactic(),
-                                      name='galactic', unit=(u.deg, u.deg))
-    else:
-        sky_frame = cf.CelestialFrame(reference_frame=coord.ICRS(),
-                                      name='icrs', unit=(u.deg, u.deg))
-
-    pipeline = [(detector_frame, det2sky), (sky_frame, None)]
-
-    return gwcs_wcs.WCS(pipeline)
-
-
-def make_imagehdu(data, wcs=None):
-    """
-    Create a FITS `~astropy.io.fits.ImageHDU` containing the input 2D
-    image.
-
-    Parameters
-    ----------
-    data : 2D array_like
-        The input 2D data.
-
-    wcs : `None` or `~astropy.wcs.WCS`, optional
-        The world coordinate system (WCS) transformation to include in
-        the output FITS header.
-
-    Returns
-    -------
-    image_hdu : `~astropy.io.fits.ImageHDU`
-        The FITS `~astropy.io.fits.ImageHDU`.
-
-    See Also
-    --------
-    make_wcs
-
-    Examples
-    --------
-    >>> from photutils.datasets import make_imagehdu, make_wcs
-    >>> shape = (100, 100)
-    >>> data = np.ones(shape)
-    >>> wcs = make_wcs(shape)
-    >>> hdu = make_imagehdu(data, wcs=wcs)
-    >>> print(hdu.data.shape)
-    (100, 100)
-    """
-    data = np.asanyarray(data)
-    if data.ndim != 2:
-        raise ValueError('data must be a 2D array')
-
-    if wcs is not None:
-        header = wcs.to_header()
-    else:
-        header = None
-
-    return fits.ImageHDU(data, header=header)
-
-
 def _define_psf_shape(psf_model, psf_shape):
     """
     Define the shape of the model to evaluate, including the
@@ -962,13 +431,11 @@ def _define_psf_shape(psf_model, psf_shape):
         model_ndim = psf_model.data.ndim
     except AttributeError:
         model_ndim = None
-        pass
 
     try:
         model_bbox = psf_model.bounding_box
     except NotImplementedError:
         model_bbox = None
-        pass
 
     if model_ndim is not None:
         if model_ndim == 3:
@@ -980,7 +447,6 @@ def _define_psf_shape(psf_model, psf_shape):
             oversampling = psf_model.oversampling
         except AttributeError:
             oversampling = 1
-            pass
         oversampling = as_pair('oversampling', oversampling)
 
         model_shape = tuple(np.array(model_shape) // oversampling)
@@ -1001,9 +467,9 @@ def _define_psf_shape(psf_model, psf_shape):
 
         if np.any(psf_shape > model_shape):
             psf_shape = tuple(np.min([model_shape, psf_shape], axis=0))
-            warnings.warn('The input psf_shape is larger than the bounding box '
-                          'size of the PSF model. The psf_shape was changed to '
-                          f'{psf_shape!r}.', AstropyUserWarning)
+            warnings.warn('The input psf_shape is larger than the bounding '
+                          'box size of the PSF model. The psf_shape was '
+                          f'changed to {psf_shape!r}.', AstropyUserWarning)
 
     return psf_shape
